==> trails/00_TRAILS.Rmd <==
---
title: "Trails"
output:
  bookdown::word_document2:
    reference_docx: ../documentation/wordtemplate.docx
    toc: true
    toc_depth: 6
  github_document:
      toc: true
      toc_depth: 6
urlcolor: blue
always_allow_html: yes
---

```{r trails-setup, include = FALSE}
knitr::opts_chunk$set(
  echo = FALSE,
  warning = FALSE,
  message = FALSE,
  fig.pos = "H",
  out.width = "65%",
  fig.height = 4,
  fig.asp = 0.6,
  dpi = 300
)

## source basic utility functions, colors for plotting, etc
source(file.path(here::here(), "R/_load_packages.R"))
source(file.path(here(), "R/_utility_functions.R"))
source(file.path(here(), "R/_global_parameters.R"))
source(file.path(here(), "R/_streetlight_credentials_parameters.R"))
source(file.path(here(), "R/_set_aesthetics.R"))
source(file.path(here(), "R/f_get_date_sequence.R"))

get_date_sequence("30-04-2022")
```

```{r fetch-trail-geography}
## set TRUE if you want to download original, unedited shapefiles
fetch_trails <- FALSE

if (fetch_trails == TRUE) {
  source(file.path(here(), "trails/01_fetch_trail_geography.R"))
}

# read saved trail geographies
trail_lines <- readRDS(file.path(here(), "data-intermediate/trails/trail-lines.RDS"))
```

```{r fetch-osm-data}
## set TRUE if you want to download OSM data within a bounding box for each trail
fetch_osm <- FALSE

if (fetch_osm == TRUE) {
  source(file.path(here(), "trails/02_fetch_osm_data.R"))
}

# load osm data & trail polygons
load(file.path(here(), "data-intermediate/trails/osm-data.rda"))
trail_polygons <- readRDS(file.path(here(), "data-intermediate/trails/trail-polygons.RDS"))
```

```{r edit-osm-segments}
## set to TRUE if you want to manually clean (add/remove) OSM segments for each trail
edit_osm_segments <- FALSE
## set to TRUE if you want to view maps of each trail & corresponding osm segments
map_issues <- FALSE

if (edit_osm_segments == TRUE) {
  source(file.path(here(), "R/_trail_geography_functions.R"))
  source_rmd(file.path(here(), "trails/03_edit_dnr_segments.Rmd"))
  source_rmd(file.path(here(), "trails/03_edit_greatermn_segments.Rmd"))
  source_rmd(file.path(here(), "trails/03_edit_metro_segments.Rmd"))
}

load(file.path(here(), "data-intermediate/trails/cleaned-trail-geography.rda"))
```

```{r create-trail-zonesets}
## set to TRUE if you want to upload new zone sets to StreetLight
create_trail_zonesets <- FALSE

if (create_trail_zonesets == TRUE) {
  source(file.path(here(), "04_create_zonesets.R"))
}
```

```{r initiate-trail-coverage}
initiate_coverage_review <- FALSE
fetch_coverage_review <- FALSE

dnr_segment_zones <- "segments-DNR_2023.06.06"
dnr_buffer_zones <- "trail_buffers_DNR_2023.06.06"

gmn_segment_zones <- "segments_greatermn_2023.07.11"
gmn_buffer_zones <- "buffers_gmn_2023.07.11"

metro_segment_zones <- "segments_metro_2023.07.11"
metro_buffer_zones <- "buffers_metro_2023.07.11"
```

```{r clean-coverage-review-data}
## set to TRUE if you want to clean the data generated by trail coverage review analyses
clean_coverage_review_data <- FALSE

if (clean_coverage_review_data == TRUE) {
  source(file.path(here(), "06_clean_coverage_review_data.R"))
}
```

```{r update-trail-metadata}
## set to TRUE to update trail metadata
update_trail_metadata <- FALSE

if (update_trail_metadata == TRUE) {
  source(file.path(here(), "trails/07_update_metadata.R"))
}
```

```{r run-core-trails}
## set run_trails to TRUE to run new LBS analyses for trails; set fetch_trails to TRUE to pull LBS data from StreetLight
## recommended to run & fetch separately
run_trails <- FALSE
fetch_trails <- FALSE

if (run_trails == TRUE) {
  source(file.path(here(), "trails_08_run_core_trails.R"))
}
```

```{r process-trail-data}
process_trail_data <- FALSE

if (process_trail_data == TRUE) {
  source(file.path(here(), "trails/09_process_trail_data.R"))
}

load(file.path(here(), "data-intermediate/processed/trail-volume.rda"))
```

```{r prep-validation-data}
prep_validation_data <- FALSE

if (prep_validation_data == TRUE) {
  # set to TRUE to run new validation analyses
  run_greatermn_validation <- FALSE
  fetch_greatermn_validation <- FALSE
  source(file.path(here(), "trails/10_prep_gmn_validation.R"))
  run_mndot_validation <- FALSE
  fetch_mndot_validation <- FALSE
  source(file.path(here(), "trails/11_prep_mndot_validation.R"))
}

gmn_validation_data <- readRDS(file.path(here(), "data-intermediate/trails/gmn-validation-data.RDS"))
load(file.path(here(), "data-intermediate/trails/mndot-validation-data.rda"))
```


```{r restrict-trail-metadata}
## restrict to trails which returned data
load(file.path(here(), "data-intermediate/trails/trail-metadata.rda"))

trail_metadata <- trail_metadata %>%
  filter(unit_id %in% annual_trail_volume$unit_id)

trail_segment_metadata <- trail_segment_metadata %>%
  filter(zone_name %in% monthly_segment_volume$zone_name)

save(trail_metadata, trail_segment_metadata,
  file = file.path(here(), "data-intermediate/trails/trail-metadata.rda")
)
```

```{r export-trail-excel}
## set to TRUE to export new excel workbook of results
export_trail_excel <- FALSE

if (export_trail_excel == TRUE) {
  source(file.path(here(), "trails/12_export_trail_excel.R"))
}
```

```{r export-trail-factsheets}
## set to TRUE if you want to create new factsheets
export_trail_factsheets <- FALSE

if (export_trail_factsheets == TRUE) {
  source(file.path(here(), "trails/13_generate_trail_factsheets.R"))
}
```

```{r create-trail-appendices}
## set to TRUE to update trail-specific appendices
create_trail_appendices <- FALSE

if (create_trail_appendices == TRUE) {
  source(file.path(here(), "trails/14_create_trail_appendices.R"))
}
```


```{r trail-documentation-setup}
cleaned_trail_segments <- unique(cleaned_trail_segments) %>%
  left_join(trail_segment_metadata) %>%
  filter(!is.na(system))

n_trails <- n_distinct(annual_trail_volume$unit_id)
n_segments <- n_distinct(monthly_segment_volume$zone_name)
n_dnr <- n_distinct(filter(annual_trail_volume, system == "DNR")$unit_id)
n_gmn <- n_distinct(filter(annual_trail_volume, system == "Greater MN")$unit_id)
n_metro <- n_distinct(filter(annual_trail_volume, system == "Metro")$unit_id)
n_months <- n_distinct(monthly_trail_volume$start_date)
n_records <- n_trails * n_months

mnoutline <- tigris::states(cb = TRUE, progress_bar = F) %>%
  filter(NAME == "Minnesota")

trail_miles <- as.numeric(sum(cleaned_trail_segments$approx_length_miles))

annual_alldays <- annual_trail_volume %>% filter(str_detect(day_type, "All"))
monthly_alldays <- monthly_trail_volume %>% filter(str_detect(day_type, "All"))
```


# Trails

Understanding trail visitation is essential for planning, programming, and investment decisions. This project used location-based services (LBS) data to analyze monthly visitation to `r n_trails` state and regional trails across Minnesota from `r analysis_period`. This data is intended to supplement, but not replace existing data used for decision making. Approximately `r prettyNum(round(sum(monthly_alldays$miles_traveled/1e6), 2), big.mark = ",")` million trail miles were traveled by visitors during the study period. In total, 2020 had the highest use across the years analyzed. Bicycle activity made up the majority of trail use, although pedestrian use made up a larger portion of use on urban trail segments and during winter months. There was considerable variation in use across segments of the same trail. 

## Methods

Aggregated and anonymized LBS data was used to estimate monthly use of `r n_trails` state and regional trails across three systems (n = `r n_dnr` Minnesota Department of Natural Resources parks (DNR State), n = `r n_gmn` Greater Minnesota Regional Parks and Trails Commission parks (Greater MN Regional), and n = `r n_metro` Twin Cities metropolitan regional parks (Metro Regional), Figure \@ref(fig:original-trails-map)). The resulting data set consists of `r n_months` monthly observations for each trail. Each observation reports estimated monthly use disaggregated by travel mode (bicycle and pedestrian).

Theoretically, the concept for measuring trail use is identical for physical trail counters (e.g., infrared or pneumatic tubes) and LBS methods. Both rely on “gates” through which visitors pass. With trail counter approaches, “gates” are physical sensors on trails. LBS data detects users passing through non-physical geographic gates. This theoretical alignment means that neither approach can identify the number of unique visitors along a trail’s entire length. For instance, visitors who do “out-and-back” trips will be counted twice (passing through the same gate once in each direction). 

Geographic data used to measure trail use were accessed via Minnesota Geospatial Commons for [DNR State](https://gisdata.mn.gov/dataset/trans-state-trails-minnesota) and [Metro Regional trails](https://gisdata.mn.gov/dataset/us-mn-state-metc-trans-regional-trails-exst-plan). Greater Minnesota Regional Parks and Trails Commission (Greater MN Regional) trail files were obtained via ArcGIS Online in November 2022. Trails were decomposed into [OpenStreetMap](https://www.openstreetmap.org/about) (OSM) trail segments for analysis (Figure \@ref(fig:camdenfig)). OSM is a widely-used mapping service used by StreetLight (the LBS data provider), private companies (i.e., Amazon, Apple, Microsoft, ESRI), and government sectors around the world. 

Each trail in the project sample consists of multiple segments split by roads or other intersecting trails/paths; however, it was beyond the scope of this project to edit or customize OSM trail segments. Trail segments less than 50 feet were removed from the analysis. This analysis contains `r n_trails` unique trails with `r prettyNum(n_segments, big.mark = ",")` distinct trail segments totaling over `r prettyNum(plyr::round_any(trail_miles, 100, f = floor), big.mark = ",")` miles (Table \@ref(tab:trail-count-tab)). Both paved and unpaved trails are represented in the data. 

LBS data was used to estimate use along each of these standardized trail segments. Bicycle and pedestrian use is considered in this project. The LBS data provider does not explicitly measure trail visitors using other transportation modes, but other modes (i.e., snowmobiles, equestrians, skiers) may be captured and classified as either bikers or walkers by the LBS data provider using their probabilistic models. Segment-level LBS estimates of trail use are conceptually identical to estimates generated by one trail counter per segment. At the segment level, we report these "counter estimates." At the trail level, we report total trail miles traveled, which is the estimated count of trail users per segment (counter estimate) multiplied by the length of the trail segment being used. When summed at the trail level, this produces a single, robust estimate of trail use. 

LBS data was obtained from StreetLight Data, Inc. and last accessed in July 2023. The StreetLight application programming interface was used to run a zone activity analysis for each trail segment’s bicycle and pedestrian use during each month of the study period. Users who traveled along any trail segment for any distance or duration were counted; no stoppage time was necessary. Devices registered to minors (under age 18) are not included in the LBS data source. 

```{r original-trails-map}
#| fig.cap = "The project sample of trails across the state of Minnesota. This map indicates original trail alignments accessed via Minnesota Geospatial Commons (DNR State, Metro Regional) or personal communication with the Executive Director of the Greater Minnesota Regional Parks and Trails Commission (Greater MN Regional).",

trail_lines %>%
  select(unit_id, geom) %>%
  filter(unit_id %in% annual_trail_volume$unit_id) %>%
  left_join(trail_metadata, by = "unit_id") %>%
  st_as_sf() %>%
  st_cast("MULTILINESTRING") %>%
  ggplot() +
  geom_sf(data = mnoutline, fill = "grey90") +
  geom_sf(aes(color = system_label)) +
  theme_void() +
  scale_color_manual(values = system_label_cols) +
  theme(
    legend.title = element_blank(),
    legend.spacing.y = unit(0, "cm")
  )
```

### Geography

Geographic data used to measure trail use was obtained from each agency. Minnesota Department of Natural Resources ([DNR state](https://gisdata.mn.gov/dataset/trans-state-trails-minnesota)) and Twin Cities metropolitan regional trails ([Metro Regional](https://gisdata.mn.gov/dataset/us-mn-state-metc-trans-regional-trails-exst-plan)) were last accessed from the Minnesota Geospatial Data Commons in July 2023. Greater Minnesota Regional Parks and Trails Commission (Greater MN Regional) trail files were obtained via ArcGIS Online in November 2022. 

In order to conduct detailed analyses, trails were decomposed into [OpenStreetMap](https://www.openstreetmap.org/about) (OSM) trail segments (Figure \@ref(fig:camdenfig)). Trail alignments were overlaid with OSM data and the intersection of these geographic data sets was used to identify relevant trail segments for each trail in the project sample. The following procedure was used to identify trail segments: 

1. Original trail alignments were buffered by 15 meters to create a single polygon representing the trail, even if the trail consists of 2 side-by-side lanes (e.g., a bicycle path adjacent to a sidewalk). Note that the ends of the original trails were *not* buffered, meaning the overall length of the trail was not affected. This step simplified trail geography in the case of side-by-side lanes and ensured that OSM segments were identified even if the alignment between OSM and agency-specific trail data was imperfect. These buffered trails, or trail polygons, were used for some LBS analyses. 

2. A "bounding box" was generated for each trail. The bounding box for a given trail is a rectangle which contains the entire extent of the trail. 

3. The [osmdata R package](https://cran.r-project.org/web/packages/osmdata/vignettes/osmdata.html) was used to query all line features within each bounding box. The original query returned many line types, such as trunk highways, bicycle paths, interstates, and residential sidewalks. 

4. Queried OSM data was restricted to OSM segments tagged for bicycle or pedestrian use. Use tags are defined within OSM and it was beyond the scope of this work to edit classifications. Other OSM segments were not discarded at this stage because, in some cases, trails consist of OSM segments *not* tagged for bicycle or pedestrian use (e.g., some trails run along abandoned railroad right-of-ways which are tagged as "abandoned" by OpenStreetMap).

5. Bicycle- and pedestrian-specific OSM segments were intersected with the buffered trail alignments described in step 1 to identify the segments most likely to be related to the trail. 

6. Manual edits were performed to add or remove OSM segments as necessary. OSM segments were typically removed if approximately 1/3 or more of the segment did not intersect with the trail alignment. Segments that consisted primarily of road intersections were generally discarded and all segments less than 50 feet in length were removed from the analysis. At this stage, OSM segments not tagged for bicycle or pedestrian use (described in step 4) were added to individual trails on a case-by-case basis to better reflect on-the-ground trail alignments.

In some cases, OSM "relations" were used to retrieve all segments related to a given trail by name (for example, querying OSM for all segments tagged "Duluth Traverse"). OSM relations were not available for the majority of trails in the project sample, hence the methods described above. Relations were used to query geographies for the Duluth Traverse (Greater MN Regional), the Luce Line (DNR State), the Paul Bunyan trail (DNR State), and the Sakatah Singing Hills trail (DNR State). 

```{r segment-medians}
median_n_segments <- median((st_drop_geometry(cleaned_trail_segments) %>%
  group_by(unit_label) %>%
  summarise(count = n(), .groups = "keep"))$count)

max_n_segments <- max((st_drop_geometry(cleaned_trail_segments) %>%
  group_by(unit_label) %>%
  summarise(count = n(), .groups = "keep"))$count)

max_segments_unit_label <- st_drop_geometry(cleaned_trail_segments) %>%
  group_by(unit_label) %>%
  summarise(count = n(), .groups = "keep") %>%
  filter(count == max_n_segments) %>%
  select(unit_label)

max_segments_trail <- trail_metadata %>%
  filter(unit_label == paste(max_segments_unit_label)) %>%
  select(long_unit_label, system)

max_segments_trail_length <- trail_metadata %>%
  filter(unit_label == paste(max_segments_unit_label)) %>%
  mutate(miles = approx_length_miles) %>%
  select(miles)

med_seg_length <- prettyNum(as.numeric(
  median(cleaned_trail_segments$approx_length_miles)
), digits = 2)
med_dnr_length <- prettyNum(as.numeric(
  median(filter(
    cleaned_trail_segments,
    str_detect(zone_name, "_DNR")
  )$approx_length_miles)
), digits = 2)
med_gmn_length <- prettyNum(as.numeric(
  median(filter(
    cleaned_trail_segments,
    str_detect(zone_name, "_Greater-MN")
  )$approx_length_miles)
), digits = 2)
med_metro_length <- prettyNum(as.numeric
(median(filter(cleaned_trail_segments, str_detect(zone_name, "_Metro-Regional"))$approx_length_miles)), digits = 2)
``` 

A total of `r prettyNum(n_segments, big.mark = ",")` distinct trails segments totaling over `r prettyNum(plyr::round_any(trail_miles, 100, f = floor), big.mark = ",")` miles were identified (Table \@ref(tab:trail-count-tab)). Please note that the identified trail segments may not reflect official trail alignments. This is for several reasons: some trails are not documented (or documented well) in OSM data; OSM data is not real-time and therefore will not reflect recent on-the-ground changes; some OSM segments follow trail alignments but extend beyond the trail's end; and short segments such as road intersections or bridge crossings were discarded. Recall that the goal of the geography editing in this project is to improve compatibility with LBS methods, not to reflect official management boundaries or alignments. 

Trails typically consist of multiple segments split by roads or other intersecting features  (Figure \@ref(fig:camdenfig)). The median number of segments per trail is `r median_n_segments`, though some long trails have significantly more segments (the maximum is `r max_n_segments` segments for the Duluth Traverse (Greater MN Regional)). 

The median segment length across the state is `r med_seg_length` miles. Because OSM segments typically break at cross-streets or similar features, OSM segments tend to be shorter in more urban or residential areas. The median segment length varies by trail system: `r med_dnr_length` miles for DNR State trails, `r med_gmn_length` miles for Greater MN Regional trails, and `r med_metro_length` miles for Metropolitan Regional trails. 

```{r trail-count-tab}
trail_summary <- st_drop_geometry(cleaned_trail_segments) %>%
  group_by(system_label) %>%
  summarise(
    Trails = n_distinct(unit_id),
    Segments = n_distinct(zone_name),
    Miles = round(sum(as.numeric(approx_length_miles))), .groups = "keep"
  )

summary_tab <- regulartable(
  trail_summary %>%
    rename(System = system_label)
) %>%
  autofit(add_w = 0, add_h = 0) %>%
  set_table_properties(layout = "autofit") %>%
  fontsize(size = 10, part = "all") %>%
  set_caption("Summary of unique trails, identified trail segments, and trail miles covered by identified trail segments across the three systems.")

summary_tab

# summary_tab %>%
#   save_as_image(file.path(here::here(), "figures/storymap/trail-summary-table.png"))
```

<br>

```{r camdenfig, out.width = "50%"}
#| fig.cap = "Example of OSM segments identified for Camden Regional Trail (Greater MN). Each colored portion of the trail represents one OSM segment. LBS methods identified seven OSM segments with an average length of 1.3 miles."

# cleaned_trail_segments %>%
#   filter(str_detect(unit_label, "Camden")) %>%
#   mapview(zcol = "osm_id", color = brewer.pal(7, "Set2"), lwd = 8, legend = FALSE) %>%
#   mapshot(file = file.path(here(), "documentation/images/camdenfig.png"))

knitr::include_graphics(file.path(here(), "documentation/images/camdenfig.png"))
```


### LBS Analyses

```{r num-trail-analyses}
n_core_za <- n_segments * 40 * 2

n_demo_analyses <- n_segments * 2 * 2

n_hourly_analyses <- n_segments * 1 * 2

n_hw_analyses <- n_trails * 2 * 2
```

Revised trail segments were used as the “zone set” for StreetLight analyses. The following StreetLight analyses were run:


On trail segments:

- Monthly Zone Activity analyses for Bicycle and Pedestrian Volume (January 2019 - April 2022). "Pass-through" trips were counted, meaning that users who traveled along any trail segment for any distance or duration were measured. Results in `r prettyNum(n_core_za, big.mark = ",")` analyses.
- Zone Activity Analysis with Traveler Attribute add-on for Bicycle and Pedestrian Volume for summer 2021 (June 1 - August 31) and the full 2021 calendar year (January 1 - December 31). Pass-through trips were counted. Results in `r prettyNum(n_demo_analyses, big.mark = ",")` analyses.
- Zone Activity Analysis with hourly time steps for Bicycle and Pedestrian Volume for summer 2021 (June 1 - August 31). Pass-through trips were counted at an hourly frequency. Results in `r prettyNum(n_hourly_analyses, big.mark = ",")` analyses.


<br>

On trail polygons:

- Zone Activity Analyses with Home-Work Location add-on for Bicycle and Pedestrian Volume for summer 2021 (June 1 - August 31) and the full 2021 calendar year (January 1 - December 31). Pass-through trips were counted. Results in `r prettyNum(n_hw_analyses, big.mark = ",")` analyses.
  
<br>

In total, `r prettyNum(sum(n_core_za + n_demo_analyses + n_hourly_analyses + n_hw_analyses), big.mark = ",")` StreetLight analyses were run for trails. StreetLight's “zone activity” analysis type is used for the entirety of the trail portion of this project. Zone activity analyses capture activity passing through, or along, each trail segment in the project sample. No stoppage time is necessary for a pass-through trip to be counted. When the LBS provider did not report data for a given trail segment in a given month (due to low use), use was assumed to be zero. 

The StreetLight application programming interface (API) was used to run each analysis. The API is accessed using [streetlightR](https://github.com/Metropolitan-Council/streetlightR), an API wrapper written in R by Metropolitan Council staff. The resulting data consists of `r prettyNum(2 * n_segments * n_months, big.mark = ",")` monthly observations (`r prettyNum(n_segments, big.mark =",")` trail segments x `r n_months` months x 2 travel modes). In most cases, data is aggregated to the trail level, resulting in `r prettyNum(2 * n_records, big.mark = ",")` monthly observations (`r prettyNum(n_trails, big.mark =",")` trails x `r n_months` months x 2 travel modes).

#### Segment-level gates

For each trail segment, the LBS provider retrieves data based on geographic gates along the segment's length (Figure \@ref(fig:gate-example)). Users must pass through the *center* gate to be counted as activity on the given segment. Segment-level LBS use estimates, therefore, are conceptually identical to estimates generated by placing physical trail counters at the "middle gate" location of each OSM segment. 

The need for users to pass through the center gate is required by StreetLight's proprietary algorithms and cannot be changed. While custom line segments can be created for use with StreetLight, creating custom lines and/or gates was beyond the scope of this project. However, such specificity may be appropriate for smaller-scale analyses focusing on fewer overall trail segments. 

Because users must pass through the center gate to be counted, it is possible that some trail users will not be counted by LBS data if they enter and exit the trail without passing through any required gates. Note that the same would be true were the geographic gate replaced with a physical trail counter. The gates underlying StreetLight's methods may affect the interpretation of segment-level data for longer trail segments. However, due to the short median segment length (`r med_seg_length` miles), it is unlikely for this limitation to affect the broader findings of this work. Further, trail segments in more walkable areas (e.g., areas where users are more likely to enter and exit trails over short distances)tend to be shorter. For example, the average identified trail segment length in Minneapolis is `r prettyNum(round(median(as.numeric(filter(cleaned_trail_segments, short_agency == "MPRB")$approx_length_mi)), 2))` miles, meaning that users would have to travel less than one tenth of a `r prettyNum(round(median(as.numeric(filter(cleaned_trail_segments, short_agency == "MPRB")$approx_length_mi)/2), 2))` miles to be missed by LBS data. 

```{r gate-example}
#| fig.cap = "An example of gates used in StreetLight's underlying methods to analyze use of OSM segments. Source: StreetLight Data, Inc. support site.",

knitr::include_graphics(file.path(here(), "documentation/images/LineZoneGates.png"))
```

### Data validation

Where possible, LBS estimates were compared to existing use estimates based on physical trail counters. Data validation was performed at individual points along trails based on the locations of physical trail counters. Due to differences in data sources and methods, perfect agreement between LBS and counter-based estimates was not expected. In particular, both LBS methods and physical trail counters may struggle to differentiate bicyclists versus pedestrians due to similar travel patterns and speed. Therefore, some misclassification was anticipated. Additionally, LBS and trail counter methods inherently represent slightly different geographies (i.e., an entire OSM trail segment versus a single physical censor) and therefore measure slightly different subsets of trail users. 

#### Greater MN data validation

```{r gmn-validation-stats}
gmn_validation_data_long <- gmn_validation_data %>%
  filter(long_count == TRUE)

gmn_cor <- cor.test(gmn_validation_data$Volume, gmn_validation_data$sadt)

gmn_cor_long <- cor.test(gmn_validation_data_long$Volume, gmn_validation_data_long$sadt)
```

LBS estimates were compared to estimates of mountain bike use on two Greater Minnesota regional trails. During the summer of 2021, use estimates were generated using a combination of trail counters and intercept surveys for the [Duluth Traverse](https://duluthmn.gov/media/13011/duluth-traverse-visitor-profile-2021.pdf) and two mountain biking trail systems in [Cook County](https://www.gmrptcommission.org/uploads/5/1/2/9/51294637/cook_county_mt_bike_visitor_profile_2021.pdf). Short-term trail counters (i.e., counters placed for less than one month) were excluded from validation because LBS data is only available at monthly increments. 

Estimates for average daily bicycle use during the summer of 2021 were compared for a segment of Short Stacker mountain biking trail at Britton Peak in Cook County and a segment of the Duluth Traverse near Spirit Mountain Recreation Area. These sites were selected for validation because counters were placed for at least one month, and the physical trail counter locations coincided with an OSM segment appropriate for LBS analysis. Note that only bicycle data was considered. 

Correlation between physical trail counter and LBS estimates were strong (r = `r prettyNum(gmn_cor_long$estimate, digits = 2)`, p < 0.001). Notably, estimates based on short-term trail counters had weaker correlation (p = `r prettyNum(gmn_cor$estimate, digits = 2)`, p < 0.001), confirming that only counters placed for a minimum of one calendar month should be considered for comparison with LBS estimates.  

LBS estimates were observed to be lower than counter-based estimates at the Cook County site, and higher than counter-based estimates at the Duluth Traverse site. These differing patterns may be explained in part by geography and nearby activity. The Cook County site is located approximately 3 miles outside of Tofte, MN (population 266 according to the [2020 US Census](https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html)) and is fairly geographically isolated. Conversely, the Duluth Traverse site is within the city limits of Duluth, MN (population 86,718 according to the [2020 US Census](https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html)) and the OSM segment used for LBS analysis runs parallel to an adjacent OSM segment belonging to the Duluth Winnipeg and Pacific railroad trail (this trail is not included in the project sample). It is possible that the LBS analysis at the Duluth Traverse site picked up some "noise", or nearby activity, on the Duluth Winnipeg and Pacific trail. However, there was insufficient evidence to confirm the precise reason for these differences. 

```{r gmn-validation-figure}
#| fig.cap = "Comparison of Greater MN counter-based estimates and LBS estimates at two long-term trail counter sites. The dashed line indicates a perfect one-to-one relationship; the solid blue line indicates the line of best fit.",
#| fig.height = 7,
#| out.width = "70%"

gmn_validation_data_long %>%
  mutate(plot_label = case_when(
    label == "Short Stacker" ~ "Cook County",
    site == "Spirit Mountain" ~ "Duluth Traverse",
    TRUE ~ NA
  )) %>%
  mutate(day_type = ifelse(day_type == "All Days", "Average Day", day_type)) %>%
  ggplot(aes(x = sadt, y = Volume)) +
  geom_smooth(method = "lm", se = FALSE, formula = y ~ x + 0, color = legacy_blue) +
  geom_point(aes(pch = day_type, fill = plot_label), size = 2) +
  geom_abline(slope = 1, intercept = 0, linetype = "dashed") +
  theme_council_open(base_size = 10) +
  theme(
    legend.position = "bottom",
    legend.box = "vertical"
  ) +
  expand_limits(x = 0, y = 0) +
  scale_shape_manual(values = c(
    "Average Day" = 21,
    "Average Weekday" = 22,
    "Average Weekend Day" = 23
  )) +
  scale_fill_manual(values = c(
    "Duluth Traverse" = legacy_green,
    "Cook County" = legacy_orange
  )) +
  labs(
    x = "Greater MN trail counter estimates", y = "LBS\ntrail use\nestimates",
    title = "Use estimates at trail counter sites \n(June - August 2021)",
    shape = "", fill = "",
    caption = "StreetLight Data, Inc; Parks and Trails Council of Minnesota. Accessed July 2023."
  ) +
  guides(fill = guide_legend(
    override.aes = list(shape = 21)
  )) +
  theme(legend.spacing.y = unit(-0.5, "cm"), )
```

#### MnDOT data validation

```{r setup-mndot-validation}
validation_sites <- filter(counter_coords, site %in% c(compare_all$site)) %>%
  left_join(zone_names) %>%
  mutate(short_site = word(Site, sep = " - ", start = 1, end = 1))
```


LBS data was additionally validated against [Minnesota Department of Transportation (MnDOT) bicycle and pedestrian counters](https://www.dot.state.mn.us/bike-ped-counting/index.html). Only permanent counters on one of the state or regional trails in the project sample were included, and sites were excluded if greater than 25% of estimates were imputed by either data source. In total, `r nrow(validation_sites)` sites were used for validation (Figure \@ref(fig:mndot-validation-sites)). Daily MnDOT estimates were converted to average monthly values for comparison with LBS estimates. Estimates for average daily bicycle and pedestrian use during each month of 2019, 2020, and 2021 were compared.

Correlation between MnDOT and LBS estimates were generally strong, with higher agreement for bicycle data (Table \@ref(tab:mndot-cor-table)). This may be explained in part by behavioral differences between bicyclists and pedestrians. For example, pedestrians may be more likely to "sidestep" a trail counter and be missed by physical trail counters than bicyclists, who are more likely to stay on designated bikeways. 
There is a noted decline in agreement, particularly for pedestrian estimates, from 2019 to 2020. This may have been caused in part by the COVID-19 pandemic if, for example, social distancing caused more pedestrians to sidestep trail counters. 

Total use estimates were compared at all sites (Figure \@ref(fig:mndot-all-sites-mode)). A strong positive relationship between trail counter and LBS estimates was observed. Visitor travel mode was also evaluated at each site. Average mode share (i.e., the percent of trail use occurring on bicycles versus on foot) was calculated at each site across all months. LBS data generally estimated a larger share of bicycle use than trail counters (Figure \@ref(fig:mndot-modeshare)). There are several potential causes for this pattern. First, as described above, some misclassification between bicycle and pedestrian activity was expected for both methods. The specific OSM segments used for analysis may also contribute to these discrepancies. For example, the OSM segment used for the West River Greenway (Metro Regional) site is tagged as a "cycleway" and runs adjacent to a separate OSM segment tagged as a "footway." This specific geography may cause the LBS provider to estimate more bicycles on the "cycleway" and more pedestrians on the "footway." However, the LBS provider's methods are proprietary, meaning there was insufficient information to determine the precise cause of this pattern. Therefore, we recommend combining bicycle and pedestrian data into a single measure of total activity for segment-level data. 

```{r}
pct_imputed_data <- pct_imputed_data %>%
  filter(zone_name %in% compare_all$zone_name)
```

Finally, estimates were compared at each site individually (Figure \@ref(fig:mndot-validation-site)). While overall correlation between LBS and counter estimates is strong, there is some variation in this relationship from site to site. It is likely that site-specific characteristics, such as geography or nearby OSM segments, contributed to this variation. Differences between the extent of OSM segments used for LBS analysis and the specific location of the physical trail counter may also play a role. Note also that at each site, between `r scales::percent(min(pct_imputed_data$mndot_pct_imputed), scale = 1)` and `r scales::percent(max(pct_imputed_data$mndot_pct_imputed), scale = 1)` of MnDOT records were imputed and up to `r scales::percent(max(pct_imputed_data$lbs_pct_missing), scale = 1)` of monthly LBS records were missing. Missing LBS data, or imputed MnDOT values, may have exacerbated discrepancies between use estimates. 

In summary, there is a strong positive correlation between MnDOT's trail counter estimates and LBS estimates. However, site-specific variation may influence the outcomes of LBS analyses. 

```{r mndot-validation-sites}
#| fig.cap = "Sites used to validate LBS against MnDOT's physical trail counters."

ggplot() +
  geom_sf(data = mnoutline, fill = "grey90", linewidth = 0) +
  geom_sf(
    data = validation_sites %>%
      mutate(short_site = str_remove_all(short_site, " Regional Trail| State Trail")),
    aes(color = short_site)
  ) +
  theme_void() +
  geom_text_repel(
    inherit.aes = FALSE,
    data = (validation_sites %>% mutate(
      short_site =
        str_remove_all(
          short_site,
          " Regional Trail| State Trail"
        )
    )),
    aes(
      x = unlist(map(validation_sites$geometry, 1)),
      y = unlist(map(validation_sites$geometry, 2)),
      label = short_site
    ),
    xlim = c(-87, NA),
    segment.color = "grey50",
    size = 4, box.padding = 1, point.padding = 0.5,
    direction = "y", hjust = 0
  ) +
  coord_sf(clip = "off") +
  guides(color = "none") +
  scale_color_brewer(palette = "Paired")
```

```{r mndot-cor-table}
cor_tab %>%
  flextable() %>%
  autofit(add_w = 0, add_h = 0) %>%
  set_table_properties(layout = "autofit") %>%
  set_caption("Pearson's correlation between LBS and trail counter data by year and mode") %>%
  bold(~ Year == "Total") %>%
  bold(j = 4)
```

```{r mndot-all-sites-mode}
#| fig.cap = "Average daily use estimates by month at six permanent MnDOT trail counter sites. The dashed line indicates a perfect one-to-one relationship; the solid blue line indicates the line of best fit. Error bars indicate standard error on MnDOT's counter estimates.",

compare_all %>%
  ggplot(aes(x = MNDOTavg, y = LBS)) +
  geom_point(size = 0.9, alpha = 0.5) +
  geom_errorbarh(aes(xmin = MNDOTavg - MNDOTse, xmax = MNDOTavg + MNDOTse),
    alpha = 0.5, lwd = 0.2
  ) +
  theme_council_open(base_size = 10) +
  scale_y_continuous(labels = scales::label_comma(), expand = c(0, 0)) +
  scale_x_continuous(labels = scales::label_comma(), expand = c(0, 0)) +
  geom_abline(slope = 1, intercept = 0, lty = 3) +
  geom_smooth(method = "lm", fill = NA, color = legacy_blue) +
  theme(legend.position = "bottom") +
  labs(
    title = "Average daily use at MnDOT counter sites \n(by month, 2019 - 2021)",
    x = "MnDOT trail counter estimate", y = "LBS \nestimate"
  )
```



```{r mndot-modeshare-setup}
mndot_modeshare_dat <- compare_all_mode %>%
  mutate(Site = str_remove_all(
    word(Site, sep = " - ", start = 1, end = 1),
    " State Trail| Regional Trail"
  )) %>%
  filter(!is.na(MNDOTavg) & !is.na(LBS)) %>%
  group_by(Site, month, year, day_type, ) %>%
  mutate(
    mndot_total = sum(MNDOTavg),
    lbs_total = sum(LBS)
  ) %>%
  group_by(Site, mode, month, year, day_type) %>%
  # mode share by month/year
  summarise(
    mndot_share = MNDOTavg / mndot_total,
    lbs_share = LBS / lbs_total, .groups = "keep"
  )
```

```{r mndot-modeshare}
#| fig.cap = "Average estimated bicycle mode share at six MnDOT trail counter sites. Error bars indicate standard error."
mndot_modeshare_dat %>%
  # average modeshare
  ungroup() %>%
  group_by(Site, mode) %>%
  summarise(
    `MnDOT counter` = mean(mndot_share),
    mndot_se = sd(mndot_share) / sqrt(n()),
    LBS = mean(lbs_share, na.rm = TRUE),
    lbs_se = sd(lbs_share, na.rm = TRUE) / sqrt(n()), .groups = "keep"
  ) %>%
  pivot_longer(cols = c(`MnDOT counter`, LBS)) %>%
  pivot_longer(cols = c(mndot_se, lbs_se), names_to = "se_name", values_to = "se") %>%
  filter(
    (name == "MnDOT counter" & se_name == "mndot_se") | (name == "LBS" & se_name == "lbs_se"),
    mode == "Bicyclist"
  ) %>%
  ggplot(aes(x = Site, y = value, fill = name, group = name)) +
  geom_col(position = "dodge", color = "black") +
  geom_linerange(aes(ymin = value - se, ymax = value + se), position = position_dodge(width = 0.9)) +
  scale_y_continuous(expand = c(0, 0)) +
  scale_fill_brewer(palette = "Paired", direction = -1) +
  theme_council_open(base_size = 10) +
  theme(legend.position = "bottom") +
  labs(
    x = "", y = "", title = "Average bicycle mode share by MnDOT site (2019 - 2021)",
    fill = "Source"
  ) +
  scale_y_continuous(labels = scales::label_percent())
```

```{r mndot-validation-site}
#| fig.cap = "Monthly use estimates at six permanent MnDOT trail counter sites by site. The dashed line indicates a perfect one-to-one relationship; the solid blue line indicates the line of best fit. Error bars indicate standard error. Note different scales for each site.",
#| out.width = "80%"

compare_all %>%
  filter(LBS != 0) %>%
  mutate(
    short_site = word(Site, sep = " - ", start = 1, end = 1),
    short_site = str_remove_all(short_site, " State Trail| Regional Trail")
  ) %>%
  ggplot(aes(x = MNDOTavg, y = LBS)) +
  geom_point(size = 0.9, alpha = 0.5) +
  geom_errorbarh(aes(xmin = MNDOTavg - MNDOTse, xmax = MNDOTavg + MNDOTse),
    alpha = 0.5, lwd = 0.2
  ) +
  facet_wrap(~short_site, scales = "free", ncol = 3, labeller = labeller(site = label_wrap_gen(25))) +
  theme_council_open(base_size = 8) +
  scale_y_continuous(labels = scales::label_comma(), expand = c(0, 0)) +
  scale_x_continuous(labels = scales::label_comma(), expand = c(0, 0)) +
  geom_abline(slope = 1, intercept = 0, lty = 3, lwd = .5) +
  geom_smooth(method = "lm", fill = NA, lwd = .5, color = legacy_blue) +
  theme(legend.position = "bottom") +
  labs(
    title = "Average daily use by MnDOT counter site (by month, 2019 - 2021)",
    shape = "Day Type", x = "MnDOT counter-based estimate", y = "LBS \nestimate"
  )
```

```{r include = FALSE}
compare_all %>%
  mutate(short_site = word(Site, sep = " - ", start = 1, end = 1)) %>%
  group_by(short_site) %>%
  summarise(
    r = cor.test(MNDOTavg, LBS)$estimate,
    p = cor.test(MNDOTavg, LBS)$p.value, .groups = "keep"
  ) %>%
  transmute(
    Site = short_site,
    r = prettyNum(r, digits = 2),
    p = ifelse(p < 0.001, "<0.001", prettyNum(p, digits = 3))
  ) %>%
  flextable() %>%
  autofit(add_w = 0, add_h = 0) %>%
  set_table_properties(layout = "autofit") %>%
  set_caption("Pearson's correlation between LBS and trail counter data site")
```


### Producing trail use estimates

Producing trail use estimates presented several unique challenges. Recall that, in the parks section of this project, park visitors were counted when they *stop* within a defined park boundary. However, given the nature of trails, we suspect that some trail users do not stop for the 5 minutes required to be counted by LBS data. Therefore, pass-through trips were counted along each trail segment in the project sample. This produces estimates similar to those generated by placing one physical trail counter on each identified trail segment. 

Whether LBS data or physical counters are used to create these segment-level estimates, summarizing the data at the trail level is non-trivial. Consider the example of Camden trail (Greater MN Regional), for which 7 trail segments were identified (Figure \@ref(fig:camdenfig)). Seven individual use estimates along the trail -- whether they be LBS or trail counter estimates -- cannot be used to identify the total number of unique users along the length of the trail. For example, if one user biked from one end of the trail to the other, they would be counted 7 times, once by each trail counter (or once per OSM trail segment). Neither method can differentiate between 1 user passing through 7 gates, or 7 users passing through 1 gate each. 

Therefore, the 7 individual estimates cannot be summed together. This would likely create an overestimate of individual trail users. Similarly, the estimates cannot reasonably be averaged. Taking the average across segments would assume that use is equal across the length of a trail, which is often untrue. In this project, we identified significant variation in use between individual segments of the same trail. 

With these challenges in mind, two primary methods for reporting trail use were explored: using the most-used trail segment as a proxy for total use, and reporting total trail miles traveled. 

There may be additional options for smaller-scale analyses focused on a smaller number of trails. For example, it may be possible to define parking areas for a trail and apply the methods described in our park methodology to estimate visitors who drive to a trailhead to access a trail. However, this method would miss any users who walk or bike onto a trail without first stopping at a designated access point. Identifying all relevant trailheads or parking areas also presents a challenge and may not be possible in all cases, particularly for trails in more walkable areas. This method would require substantial on-the-ground knowledge about individuals trails and is, therefore, beyond the scope of this project. Similarly, it may be possible to augment LBS data with other data sources, such as intercept survey results, to create an estimate of individual trail users. The methods explored in this report, however, rely solely on LBS data. 

#### Most-used segment approach (counter estimates)

One option for reporting trail data is to consider the most-used trail segment on a given trail as a proxy for total trail use. The rationale behind this approach is that the most-used trail segment is likely where a single physical trail counter would be placed, were physical counters being used. Following this reasoning, we refer to these estimates as "counter estimates" throughout the project. Most-used segments were determined on a monthly basis to better reflect changing use patterns throughout the year. 

```{r setup-summer-metro-comp}
metc_trailuse_21 <- 17965.4 * 1000

annual_counter_ests <- annual_trail_volume %>%
  filter(str_detect(day_type, "All")) %>%
  group_by(system, year) %>%
  summarise(total_counter_estimate = sum(counter_estimate), .groups = "keep")

lbs_counter_trailuse_21 <- annual_counter_ests %>%
  ungroup() %>%
  filter(system == "Metro" & year == 2021) %>%
  select(total_counter_estimate) %>%
  as.numeric()

load(file.path(here(), "data-raw/metro/estimates_summer.rda"))
load(file.path(here(), "data-raw/metro/park_id_name_index.rda"))

summer_trail_counts <- estimates_summer$d2021_summer_estimates %>%
  left_join(park_id_name_index) %>%
  filter(str_detect(Park_name, "Trail")) %>%
  mutate(trail_name = str_remove_all(Park_name, " Regional Trail")) %>%
  # get one record per trail
  group_by(Park_id) %>%
  slice(1) %>%
  mutate(
    short_agency = case_when(
      str_detect(Agency_name, "Three") ~ "Three Rivers",
      str_detect(Agency_name, "Minneapolis") ~ "MPRB",
      TRUE ~ Agency_name
    ),
    unit_label = paste0(trail_name, ", ", short_agency),
    # force unit_labels to match
    unit_label = case_when(
      unit_label == "Mississippi River Greenway, Dakota County" ~ "Mississippi River, Dakota County",
      unit_label == "Minnesota River Bluffs LRT, Three Rivers" ~ "MN River Bluffs, Three Rivers",
      unit_label == "Samuel H. Morgan, Saint Paul" ~ "Sam Morgan, Saint Paul",
      unit_label == "River to River Greenway, Dakota County" ~ "River to River, Dakota County",
      unit_label == "Minnesota River Greenway, Dakota County" ~ "Minnesota River, Dakota County",
      unit_label == "Nokomis-Minnesota River, Three Rivers" ~ "Nokomis - Minnesota River, Three Rivers",
      unit_label == "Nokomis-Minnesota River, Bloomington" ~ "Nokomis - Minnesota River, Bloomington",
      unit_label == "West Mississippi River, Three Rivers" ~ "West Mississippi, Three Rivers",
      TRUE ~ unit_label
    )
  )

summer_metro_comp <- seasonal_trail_volume %>%
  filter(
    str_detect(system, "Metro"),
    year == 2021,
    season == "Summer",
    str_detect(day_type, "All")
  ) %>%
  left_join(summer_trail_counts, by = c("unit_label", "short_agency")) %>%
  select(unit_id, unit_label, year, LBS = counter_estimate, count_est = Estimate)

summer_metro_cor <- cor.test(summer_metro_comp$LBS, summer_metro_comp$count_est)

pct_diff_by_unit <- summer_metro_comp %>%
  filter(!is.na(count_est) & !is.na(LBS)) %>%
  group_by(unit_label) %>%
  summarise(
    count_est = count_est,
    LBS = LBS,
    diff = count_est - LBS,
    pct = diff / count_est,
    abs_pct = abs(pct), .groups = "drop"
  )
```

LBS counter estimates were found to be reasonable in some cases. For example, the [Metropolitan Council's annual park use estimates](https://metrocouncil.org/Parks/Research/Annual-Use-Estimates.aspx) report `r prettyNum(metc_trailuse_21, big.mark = ",")` total trail visits in 2021. The LBS most-used segment approach estimated `r prettyNum(lbs_counter_trailuse_21, big.mark = ",")` total visits in the same year, a difference of only `r scales::percent((lbs_counter_trailuse_21 - metc_trailuse_21)/lbs_counter_trailuse_21)`. At this high level of summary, LBS counter estimates may be sufficient. 

LBS counter estimates were additionally compared to unit-level summer 2021 use estimates for Metro Regional trails (Figure \@ref(fig:summer-metro-counter-comp)). The existing estimates are based on manual counts performed during the summer (June - August 2021). Across all Metro Regional trails for which data was available (n = `r nrow(pct_diff_by_unit)`), agreement between LBS counter estimates and manual count-based estimates was moderate (r = `r prettyNum(round(summer_metro_cor$estimate, digits = 2))`, p < 0.001). Results differed for individual trails and, overall, the most-used segment approach underestimated trail use compared to the count-based use estimates. At individual trails, the difference between LBS counter estimates and count-based estimates ranged from `r scales::percent(min(pct_diff_by_unit$abs_pct))` (`r filter(pct_diff_by_unit, abs_pct == min(abs_pct))$unit_label`) to `r scales::percent(max(pct_diff_by_unit$abs_pct))` (`r filter(pct_diff_by_unit, abs_pct == max(abs_pct))$unit_label`). 

This approach is sensitive to the specific OSM segments used to represent a trail. A short segment, such as a bridge or major road crossing, may capture a significant number of "incidental users" and appear to be the most-used segment of a trail. If such a segment were not included in the analysis for any reason (i.e., the OSM segment was not detected using the procedure described above, or the segment was less than 50ft in length), a given trail may have a very different use estimate. This method may also misrepresent trails which are very high-use across their entire length, or which have significant variation between their most- and least-used segments. Additionally, by reporting only one value, this method inherently limits the amount of detail included in use estimates and further analysis. Because of these challenges, and the relatively weak correlation with existing estimates observed, this approach is not the preferred method for this project. 

Note that this method may be appropriate for specific trails, or specific research questions. Therefore, the LBS counter estimate is provided at the segment-level in the final project data. LBS counter estimates are also used throughout this project when segment-level data is presented. 

```{r summer-metro-counter-comp}
#| fig.cap = "Comparison of count-based use estimates and LBS counter estimates at Metro Regional trails. Data for summer of 2021 (June - August) is compared. The dashed line indicates a perfect one-to-one relationship; the solid blue line indicates the line of best fit. Trails with the largest discrepancies are labeled.",
#| out.width = "70%",
#| fig.height = 5

summer_metro_comp %>%
  mutate(label = if_else(abs(count_est - LBS) > 140000, unit_label, NA_character_)) %>%
  ggplot(aes(x = count_est, y = LBS, label = label)) +
  geom_point(
    alpha = .3
  ) +
  geom_smooth(method = "lm", formula = y ~ x + 0, se = FALSE, color = legacy_blue) +
  geom_abline(slope = 1, intercept = 0, linetype = 2) +
  theme_council_open(base_size = 10) +
  scale_x_continuous(labels = scales::label_comma(), expand = expansion(mult = c(0, 0.01))) +
  scale_y_continuous(labels = scales::label_comma(), expand = c(0, 0)) +
  labs(
    title = "Use estimates at Metro Regional trails",
    x = "Count-based use estimate", y = "LBS counter \nestimate",
    caption = paste0("Source: StreetLight Data, Inc; Metropolitan Council Use Estimates. Accessed July 2023.")
  ) +
  geom_text_repel(
    min.segment.length = 0,
    box.padding = 0.35,
    point.padding = 0.5,
    segment.color = "grey50",
    size = 3,
    lineheight = 0.8,
    xlim = c(300000, NA),
    max.overlaps = 4,
    nudge_y = 50000
  )
```

```{r include = FALSE}
cvt_renee <- data.frame(
  year = c(2019, 2020, 2021),
  est = c(89664, 171940, 125854)
)

mes_renee <- data.frame(
  year = c(2019, 2020),
  est = c(223351, 248219)
)

cvt_lbs <- annual_trail_volume %>%
  ungroup() %>%
  filter(
    str_detect(day_type, "All"),
    str_detect(unit_label, "Cannon Valley"),
    year != 2022
  ) %>%
  select(year, counter_estimate)

mes_lbs <- annual_trail_volume %>%
  ungroup() %>%
  filter(
    str_detect(day_type, "All"),
    str_detect(unit_label, "Mesabi"),
    year %in% c(2019, 2020)
  ) %>%
  mutate(unit_label = "Mesabi Trail, all sections") %>%
  group_by(year) %>%
  summarise(counter_estimate = max(counter_estimate), .groups = "keep")

gmn_most_used_comp <- left_join(cvt_renee, cvt_lbs) %>%
  mutate(unit_label = "Cannon Valley") %>%
  bind_rows(left_join(mes_renee, mes_lbs) %>%
    mutate(unit_label = "Mesabi Trail, all sections"))

gmn_most_used_comp %>%
  ggplot(aes(x = est, y = counter_estimate)) +
  geom_point() +
  geom_abline(slope = 1, intercept = 0) +
  geom_smooth(method = "lm", se = FALSE)
```

#### Trail miles traveled

```{r setup-tmt}
cor_user_travel <- annual_alldays %>%
  mutate(
    max_use = counter_estimate / 1e6,
    miles_used = miles_traveled / 1e6
  )

cor_eq <- lm(miles_used ~ 0 + max_use, data = cor_user_travel)

cor_user_travel_unit <- cor_user_travel %>%
  group_by(unit_label, unit_id, year) %>%
  summarise(miles_per_user = miles_used / max_use, .groups = "keep") %>%
  filter(!is.nan(miles_per_user)) %>%
  left_join(trail_metadata)

tmt_counter_cor <- cor.test(cor_user_travel$max_use, cor_user_travel$miles_used)
```

A key goal of this analysis was to create a measure of trail use which performs similarly across trails and is robust to small changes in identified OSM segments. To achieve these goals, we propose trail miles traveled (TMT) as the primary method for reporting trail-level data. 

Trail miles traveled is the estimated count of users on a given segment (i.e., the LBS counter estimate described above) multiplied by the length of said trail segment. As a measure of trail use, trail miles traveled is conceptually similar to vehicle miles traveled (VMT), a measure of vehicle traffic [used by the US Department of Transportation](https://www.transportation.gov/mission/health/vmt-capita). When summed at the trail level, TMT provides a single, robust estimate of trail use. 

Naturally, estimated trail miles traveled will not directly correlate to existing use estimates. This is because TMT represents a trail-level summary of use, while existing estimates represent individual trail users. Therefore, to build confidence in the proposed trail miles traveled metric, annual TMT was compared to the LBS counter estimate described above (Figure \@ref(fig:tmt-max-fig)). The purpose of this comparison was to confirm that TMT performs similarly for trails across the project sample. On average, we found that `r round(cor_eq$coefficients[1], 1)` trail miles were traveled for every user counted at the most-used segment of a given trail. While there was some variation in the relationship between LBS counter estimates and estimated TMT, the variation can largely be explained. 

For example, the Minneapolis Chain of Lakes trail (Metro Regional) showed more trail miles traveled relative to users counted at its most-used segment. This trail is located in a densely populated area and encircles the entirety of Lake Harriet and Bde Maka Ska and the western shore of Cedar Lake over `r round(trail_metadata %>% filter(str_detect(unit_label, "Minneapolis Chain of Lakes")) %>% transmute(approx_length_miles = as.numeric(approx_length_miles)), 2)` miles. In most places, the trail is made up of parallel walking and biking paths. The most used segment is along the western shore of Lake Harriet. Due to the popularity of circumnavigating a single lake, it logically follows that total trail miles traveled may be more than `r round(cor_eq$coefficients[1], 1)` times the use at this segment. Indeed, trail miles traveled was `r round(filter(cor_user_travel_unit, year == 2020 & unit_label == "Minneapolis Chain of Lakes, MPRB")$miles_per_user, 1)` times greater than the maximum segment use for this trail (in the year 2020).

The Paul Bunyan trail (DNR State) is another example of a trail with more trail miles traveled relative to its LBS counter estimate. This is, in large part, due to the length of the trail: `r round(trail_metadata %>% filter(str_detect(unit_label, "Paul Bunyan")) %>% transmute(approx_length_miles = as.numeric(approx_length_miles)), 1)` miles. The most-used segment of this trail was less than one mile long and ran along a residential street in Bemidji, MN. Naturally, a trail of this length will have more than one highly-used segment, meaning a single LBS counter estimate is likely to underestimate use along the entire trail. Indeed, in 2020, TMT was `r round(filter(cor_user_travel_unit, year == 2020 & unit_label == "Paul Bunyan")$miles_per_user, 1)` greater than use on the single highest-use segment in Bemidji. 

Conversely, the Central Mississippi Riverfront trail (Metro Regional) showed fewer trail miles traveled relative to users counted at its most-used segment. This trail spans both banks of the Mississippi River in downtown Minneapolis, including two relatively short but high-use areas (the Hennepin Avenue bridge across the Mississippi and a trail segment bordering Mill Ruins Park and the Stone Arch Bridge). A large number of visitors may use either bridge for utilitarian purposes or attend one of several popular festivals held in Mill Ruins Park. In either of those cases, visitors may use only a short segment of the trail. In this case, TMT may be less than `r round(cor_eq$coefficients[1], 1)` times the LBS counter estimate. In 2020, trail miles traveled was `r round(filter(cor_user_travel_unit, year == 2020 & unit_label == "Central Mississippi Riverfront, MPRB")$miles_per_user, 1)` times the maximum segment use, in line with this hypothesis.

Cedar Lake trail (Metro Regional, Three Rivers section) also showed fewer trail miles traveled relative to users counted at its most used segment. The explanation in this case is simple: in 2019, large portions of the trail were closed for light rail construction. While the open segments continue to receive high use, use on other segments is near zero, causing trail miles traveled to be only `r round(filter(cor_user_travel_unit, year == 2019 & unit_label == "Cedar Lake, Three Rivers")$miles_per_user, 1)` times the maximum segment use in 2019. This pattern may extend to other trails with small areas that have much higher use than the rest of the trail.  

```{r pb-robustness-example}
pb_ests1 <- monthly_segment_volume %>%
  select(-modeshare_bike_miles, -modeshare_ped_miles) %>%
  # summarize at trail level
  group_by(unit_id, start_date, month, year, day_type) %>%
  summarise(
    counter_estimate = max(counter_estimate),
    bike_counter_estimate = max(bike_counter_estimate),
    ped_counter_estimate = max(ped_counter_estimate),
    miles_traveled = sum(miles_traveled),
    bike_miles_traveled = sum(bike_miles_traveled),
    ped_miles_traveled = sum(ped_miles_traveled), .groups = "keep"
  ) %>%
  mutate(
    modeshare_bike_miles = bike_miles_traveled / miles_traveled,
    modeshare_ped_miles = ped_miles_traveled / miles_traveled
  ) %>%
  mutate(
    modeshare_bike_miles = ifelse(is.nan(modeshare_bike_miles), NA, modeshare_bike_miles),
    modeshare_ped_miles = ifelse(is.nan(modeshare_ped_miles), NA, modeshare_ped_miles)
  ) %>%
  replace_na(list(modeshare_bike_miles = 0, modeshare_ped_miles = 0)) %>%
  left_join(trail_metadata) %>%
  filter(
    unit_label == "Paul Bunyan",
    year == 2021,
    str_detect(day_type, "All")
  ) %>%
  group_by(year) %>%
  summarise(
    annual_count_estimate = sum(counter_estimate),
    annual_tmt = sum(miles_traveled), .groups = "keep"
  )

pb_ests2 <- monthly_segment_volume %>%
  filter(as.numeric(approx_length_meters) > 200) %>%
  select(-modeshare_bike_miles, -modeshare_ped_miles) %>%
  # summarize at trail level
  group_by(unit_id, start_date, month, year, day_type) %>%
  summarise(
    counter_estimate = max(counter_estimate),
    bike_counter_estimate = max(bike_counter_estimate),
    ped_counter_estimate = max(ped_counter_estimate),
    miles_traveled = sum(miles_traveled),
    bike_miles_traveled = sum(bike_miles_traveled),
    ped_miles_traveled = sum(ped_miles_traveled), .groups = "keep"
  ) %>%
  mutate(
    modeshare_bike_miles = bike_miles_traveled / miles_traveled,
    modeshare_ped_miles = ped_miles_traveled / miles_traveled
  ) %>%
  mutate(
    modeshare_bike_miles = ifelse(is.nan(modeshare_bike_miles), NA, modeshare_bike_miles),
    modeshare_ped_miles = ifelse(is.nan(modeshare_ped_miles), NA, modeshare_ped_miles)
  ) %>%
  replace_na(list(modeshare_bike_miles = 0, modeshare_ped_miles = 0)) %>%
  left_join(trail_metadata) %>%
  filter(
    unit_label == "Paul Bunyan",
    year == 2021,
    str_detect(day_type, "All")
  ) %>%
  group_by(year) %>%
  summarise(
    annual_count_estimate = sum(counter_estimate),
    annual_tmt = sum(miles_traveled), .groups = "keep"
  )
```


Trail miles traveled is more robust to changes in identified OSM segments than the LBS counter estimates described above. For example, the 2021 LBS counter estimate for the Paul Bunyan trail (DNR State) was `r prettyNum(filter(annual_alldays, unit_label == "Paul Bunyan" & year == 2021)$counter_estimate, big.mark = ",")`. If, for example, all OSM segments shorter than 200m were discarded from the LBS analysis, this estimate would decrease to `r prettyNum(pb_ests2$annual_count_estimate, big.mark = ",")`, a change of -`r scales::percent((filter(annual_alldays, unit_label == "Paul Bunyan" & year == 2021)$counter_estimate - pb_ests2$annual_count_estimate) / filter(annual_alldays, unit_label == "Paul Bunyan" & year == 2021)$counter_estimate)`. Conversely, the TMT estimate would only decrease by `r scales::percent((pb_ests1$annual_tmt - pb_ests2$annual_tmt) / pb_ests1$annual_tmt)`. By combining data across all identified OSM segments, the TMT metric reduces the potential impact of adding or removing a single trail segment. Similarly, this method reduces the impact of site-specific characteristics (described in the data validation portion of this report) that may affect LBS estimates.

Unless otherwise specified, trail use is reported as trail miles traveled in this project. Bicycle and pedestrian use is typically combined into a single value, but disaggregated data (i.e., bicycle miles traveled and pedestrian miles traveled) is provided in the final project data. When visitor mode share is reported, it is calculated based on the share of bicycle and pedestrian miles traveled. 

```{r tmt-max-fig}
#| fig.cap = "Comparison of annual trail miles traveled and annual maximum segment use (LBS counter estimate) for all trails across the three complete years in this study (2019, 2020, 2021). The solid line indicates the line of best fit. Trails with the largest discrepancies are labeled.",
#| out.width = "70%",
#| fig.height = 5

cor_user_travel %>%
  mutate(
    calc = max_use * cor_eq$coefficients[1],
    label = if_else(abs(calc - miles_used) > 2, paste(unit_label, year), NA_character_),
    label = ifelse(unit_label == "Paul Bunyan" & year == 2020, paste(unit_label, year), label),
    label = str_remove(label, " MPRB")
  ) %>%
  ggplot(aes(x = max_use, y = miles_used, label = label)) +
  geom_point(alpha = .3) +
  geom_smooth(aes(group = 1), method = "lm", color = legacy_blue, formula = y ~ x + 0, se = FALSE) +
  theme_council_open(base_size = 10) +
  theme(legend.position = "none") +
  labs(
    shape = "", color = "", y = "Total trail\nmiles traveled\n", x = "LBS counter estimate",
    title = "LBS counter estimates and trail miles traveled",
    caption = fig_caption
  ) +
  geom_text_repel(aes(label = label),
    min.segment.length = 0,
    box.padding = 0.35,
    point.padding = 0.5,
    segment.color = "grey50",
    size = 3, color = "black",
    nudge_y = 1,
    max.overlaps = 3
  )

# ggsave(file.path(here::here(), "figures/storymap/tmt_vs_max.png"))
```


## Results


### Trail use increased in 2020

Across the state, 2020 saw a distinct increase in trail use (`r prettyNum(round(sum(monthly_alldays$miles_traveled[monthly_alldays$year == 2020]/1e6), 2), big.mark = ",")` million trail miles traveled in 2020 across three systems compared to `r prettyNum(round(sum(monthly_alldays$miles_traveled[monthly_alldays$year == 2019]/1e6), 2), big.mark = ",")` and `r prettyNum(round(sum(monthly_alldays$miles_traveled[monthly_alldays$year == 2021]/1e6), 2), big.mark = ",")` million miles traveled in 2019 and 2021, respectively). This increase was likely driven by behavior changes caused by the COVID-19 pandemic. Trail use in 2021 largely returned to pre-pandemic levels, though use was higher in 2021 than 2019. The three systems saw proportionally similar changes in trail use. 

At the unit level, not all trails had increased use in 2020. Year 2020 had the highest use for `r annual_alldays %>% group_by(unit_label) %>% filter(miles_traveled == max(miles_traveled)) %>% ungroup() %>% count(year) %>% .[2,2]` trail units, but `r annual_alldays %>% group_by(unit_label) %>% filter(miles_traveled == max(miles_traveled)) %>% ungroup() %>% count(year) %>% .[1,2]` and `r annual_alldays %>% group_by(unit_label) %>% filter(miles_traveled == max(miles_traveled)) %>% ungroup() %>% count(year) %>% .[3,2]` units had the highest use in years 2019 and 2021, respectively.


```{r annual-trail-miles}
#| fig.cap = "Total estimated trail use (trail miles traveled) by year. Bars show the system-wide total trail use across the three complete years in this study (2019, 2020, 2021). Bar color indicates trail system. Individual trail units may show annual visitation trends that differ from the overall system trend.",


annual_alldays %>%
  filter(year != 2022) %>%
  group_by(year, system_label) %>%
  summarise(miles = sum(miles_traveled) / 1e6) %>%
  ggplot(aes(x = year, y = miles, fill = system_label)) +
  geom_bar(stat = "identity", color = "black") +
  geom_label(aes(label = format(round(miles, 2), nsmall = 2), y = rep(c(6, 3.3, 30), 3)), fill = "white", size = 2.5) +
  labs(
    title = "Annual trail use (millions of miles traveled)", y = "", x = "",
    caption = fig_caption
  ) +
  theme_council_open(base_size = 10) +
  coord_cartesian(clip = "off") +
  scale_fill_manual(values = system_label_cols) +
  theme(
    legend.position = "none",
    axis.title.y = element_blank(),
    axis.title.x = element_blank(),
    plot.title = element_text(size = 10)
  ) +
  facet_wrap(~system_label, scales = "free_y", nrow = 1) +
  scale_y_continuous(expand = c(0, 0))
# ggsave(file.path(here::here(), "figures/storymap/annual-trail-use.png"))
```

### Segment-level data provides detailed insight

Trails connect Minnesotans to a wide range of land uses, essential services, cities, and other attractions. In addition to the recreation value of trails, trails are essential components of transportation networks, connect park systems, and advance natural resource conservation efforts across the state. Due to the variety of landscapes and diverse services trails provide, not all segments of trails are used equally. 

While trail use estimates are often limited to several short-term trail counters sited in strategic locations, LBS methods can estimate use across a trail’s entire length and reveal spatial and temporal variation in trail use.  

For example, consider a trail with several temporary closures along its length (e.g., portions of Cedar Lake trail (Metro Regional) closed in  May 2019 for light rail construction. A single trail counter might indicate unchanged trail use or trail use falling to zero, depending on the counter’s specific location (a single “gate”). Segment-level LBS estimates reflect changes to all trail segments – both those which see declines in visitation and those where visitation remains high – better reflecting on-the-ground conditions (Figure \@ref(fig:monthly-trail-ts)). 

Similarly, trails that are primarily rural with some segments passing through population centers may have large differences between their most used trail segments. For example, the most-used segment of Central Lakes trail (DNR State) is directly northwest of Alexandria, MN (Figure \@ref(fig:centralfig)). Alexandria is the largest population center directly along the trail with a population of approximately 14,000 ([2020 US Census](https://www.census.gov/data/tables/time-series/demo/popest/2020s-total-cities-and-towns.html)). The population and activity associated with Alexandria contributes to the large gap between most- and least-used segments on this trail (Figure \@ref(fig:monthly-trail-ts)).

<!-- This data can be explored for each trail and trail segment. In the interactive figure below (Figure 4.5), select a trail of interest to see a time series of trail use from January 2019 - April 2022, a map of activity by trail segment, a time series of mode share, and a summary table. Hover over any element and/or zoom in for additional detail.  -->


```{r monthly-trail-ts}
#| fig.cap = 'Estimated monthly segment-level trail use at two trails. For Cedar Lake Regional Trail (top panel), a segment which closed for construction in 2019 is highlighted. For Central Lakes State Trail (bottom panel), the most- and least-used segments are each highlighted. Each line corresponds to an individual OSM trail segment, therefore the values shown represent LBS counter estimates (rather than trail miles traveled). ',
#| fig.asp = 1.2,
#| out.width = "75%"

a <- monthly_segment_volume %>%
  filter(
    unit_label == "Cedar Lake, Three Rivers",
    str_detect(day_type, "All")
  ) %>%
  mutate(
    start_date = as.Date(start_date, format = "%m/%d/%Y"),
    flag = ifelse(osm_id == "706678932", TRUE, FALSE),
    label = ifelse(start_date == "2019-05-01" & osm_id == "706678932", "Segment closed", NA_character_),
    # force drawing order
    osm_id = factor(osm_id, levels = c(unique(filter(
      monthly_segment_volume,
      unit_label == "Cedar Lake, Three Rivers" &
        str_detect(day_type, "All") &
        osm_id != 706678932
    )$osm_id), 706678932), ordered = TRUE)
  ) %>%
  ggplot(aes(x = start_date, y = counter_estimate, group = osm_id, color = flag, alpha = flag, size = flag)) +
  geom_line() +
  scale_color_manual(values = c(
    "TRUE" = "black",
    "FALSE" = "grey50"
  )) +
  scale_alpha_manual(values = c(
    "TRUE" = 1,
    "FALSE" = 0.5
  )) +
  scale_size_manual(values = c(
    "TRUE" = 1,
    "FALSE" = 0.5
  )) +
  scale_x_date(date_breaks = "6 months", date_labels = "%b '%y", expand = c(0, 0), limits = c(as.Date("2019-01-01"), as.Date("2022-04-30"))) +
  scale_y_continuous(labels = scales::label_comma(), expand = expansion(mult = c(0.0, 0.1))) +
  theme_council_open(base_size = 10) +
  theme(
    legend.position = "none",
    panel.grid.major.x = element_line(linetype = 3),
    axis.title.x = element_blank()
  ) +
  labs(
    y = "",
    title = "Cedar Lake Regional Trail (Three Rivers)"
  ) +
  geom_text_repel(aes(label = label),
    segment.color = "grey50", min.segment.length = 0,
    size = 3, box.padding = 1, point.padding = 0.2,
    segment.square = TRUE, nudge_x = 150, nudge_y = 1500,
    lineheight = 0.8, max.overlaps = 2
  )

b <- monthly_segment_volume %>%
  filter(
    unit_label == "Central Lakes",
    str_detect(day_type, "All")
  ) %>%
  mutate(
    start_date = as.Date(start_date, format = "%m/%d/%Y"),
    flag = ifelse(osm_id %in% c(769784624, 769784620), TRUE, FALSE),
    label = case_when(
      osm_id == 769784620 ~ "Least-used segment",
      osm_id == 769784624 ~ "Most-used segment",
      TRUE ~ NA_character_
    ),
    text_label = case_when(
      start_date == as.Date("2019-09-01") & osm_id == 769784620 ~ "Least-used \nsegment",
      start_date == as.Date("2019-09-01") & osm_id == 769784624 ~ "Most-used \nsegment",
      TRUE ~ NA_character_
    ),
    osm_id = factor(osm_id, levels = c(unique(filter(
      monthly_segment_volume,
      unit_label == "Central Lakes" &
        str_detect(day_type, "All") &
        !osm_id %in% c(769784624, 769784620)
    )$osm_id), 769784624, 769784620), ordered = TRUE)
  ) %>%
  ggplot(aes(x = start_date, y = counter_estimate, group = osm_id, color = label, alpha = flag, size = flag)) +
  geom_line() +
  scale_color_manual(values = c(
    "Least-used segment" = legacy_blue,
    "Most-used segment" = legacy_green
  )) +
  scale_alpha_manual(values = c(
    "TRUE" = 1,
    "FALSE" = 0.5
  )) +
  scale_size_manual(values = c(
    "TRUE" = 1,
    "FALSE" = 0.5
  )) +
  scale_x_date(date_breaks = "6 months", date_labels = "%b '%y", expand = c(0, 0), limits = c(as.Date("2019-01-01"), as.Date("2022-04-30"))) +
  scale_y_continuous(labels = scales::label_comma(), expand = expansion(mult = c(0.001, 0.1))) +
  theme_council_open(base_size = 10) +
  theme(
    legend.position = "none",
    panel.grid.major.x = element_line(linetype = 3),
    axis.title.x = element_blank()
  ) +
  labs(
    y = "",
    title = "Central Lakes State Trail",
    caption = fig_caption
  ) +
  geom_label_repel(aes(label = text_label),
    min.segment.length = 0,
    size = 3, box.padding = 2, point.padding = 1,
    segment.square = TRUE,
    lineheight = 0.8, max.overlaps = 3,
    nudge_x = -60
  )

segment_title <- ggdraw() +
  draw_label(
    "Monthly segment-level use estimates",
    x = 0,
    hjust = 0
  ) +
  theme(
    plot.margin = margin(0, 0, 0, 7)
  )

plot_grid(segment_title, a, b,
  rel_heights = c(0.1, 1, 1),
  ncol = 1
)
```



```{r centralfig}
#| fig.cap = "Most- and least-used trail segments on Central Lakes State Trail. Note that the most-used trail segment is directly northwest of Alexandria, MN. ",
#| out.width = "80%"

# cleaned_trail_segments %>%
#   filter(unit_label == "Central Lakes") %>%
#   mutate(label = case_when(
#            osm_id == 769784620 ~ "Least-used segment",
#            osm_id == 769784624 ~ "Most-used segment",
#            TRUE ~ NA_character_
#          )) %>%
#   mapview(zcol = "label", color = c("#003865", "#2DA44A"), lwd = 6, legend = FALSE) %>%
#   mapshot(vwidth = 800, vheight = 600,
#           file = file.path(here(), "documentation/images/centrallakesfig.png"))

# knitr::include_graphics(file.path(here(), "documentation/images/centrallakesfig.png"))

gg_base <- ggplot() +
  labs(title = "Most- and least-used segments on Central Lakes State Trail") +
  theme_void() +
  theme(title = element_text(hjust = 0.5, size = 8))

ggdraw(gg_base) +
  draw_image(file.path(here(), "documentation/images/centrallakesfig.png"),
    y = -0.08, x = -0.1
  )
```

#### Visitor travel mode differs by trail segment

In addition to overall use, visitor travel mode was found to vary by trail segment. Like overall use, we observe a relationship between visitor mode share and population centers. Trail segments located near walkable areas, such as downtowns or residential neighborhoods, may receive higher pedestrian use than other segments on the same trail. In the case of Central Lakes (DNR State), pedestrian activity makes up a much larger share of trail use near Alexandria and in Osakis at the southeastern end of the trail (Figure \@ref(fig:central-mode-fig)). We observe that this change is caused by an increase in pedestrian use in these areas, rather than a decrease in bicycle activity. 

```{r central-mode-fig}
#| fig.cap = "Visitor mode share by trail segment on Central Lakes State Trail. Brighter colors indicate a higher portion of pedestrian trail use. Note that trail segments closer to population centers (e.g., northwest of Alexandria, MN) or in residential areas (e.g., southeastern end of the trail in Osakis, MN) have higher portions of pedestrian use.",
#| out.width = "80%"

# cleaned_trail_segments %>%
#   filter(unit_label == "Central Lakes") %>%
#   left_join(monthly_segment_volume %>%
#               filter(unit_label == "Central Lakes", year == 2021)  %>%
#               group_by(unit_label, osm_id) %>%
#               summarise(bike = sum(bike_counter_estimate),
#                         ped = sum(ped_counter_estimate), .groups = "keep") %>%
#               mutate(bike_modeshare = 100*bike/(bike+ped),
#                      `Pedestrian mode share (%)` = 100*ped/(bike+ped))) %>%
#   filter(!is.na(`Pedestrian mode share (%)`)) %>%
#   mapview(zcol = "Pedestrian mode share (%)", lwd = 6) %>%
#   mapshot(vwidth = 800, vheight = 600,
#           file = file.path(here(), "documentation/images/centrallakes_mode_fig.png"))

# knitr::include_graphics(file.path(here(), "documentation/images/centrallakes_mode_fig.png"))

gg_base2 <- ggplot() +
  labs(title = "Pedestrian mode share by segment on Central Lakes State Trail") +
  theme_void() +
  theme(title = element_text(hjust = 0.5, size = 8))

ggdraw(gg_base2) +
  draw_image(file.path(here(), "documentation/images/centrallakes_mode_fig.png"),
    y = -0.08, x = -0.1
  )
```

### Trails are used in all seasons

Using LBS data to estimate trail use over a year shows that Minnesota’s trails are used in all seasons, although to a lesser extent than Minnesota’s parks. Trail use generally builds throughout the spring, peaks in the summer, and tapers off in the fall (Figure \@ref(fig:seasonal-fig-trail)). At the trail level, many segments which retain usage into the winter are within population centers or maintained (plowed) during the winter (e.g., portions of Brown’s Creek State Trail), highlighting the year-round importance of trails to the transportation network and for recreation benefits. 

```{r seasonal-fig-trail}
#| fig.cap = "Monthly trail use as a percent of total annual use for each system. Points show the average monthly use across the three complete years in this study (2019, 2020, 2021). Point color indicates park system. Individual trail units may show trends that differ from the overall system trend.",

(monthly_alldays) %>%
  filter(year != 2022) %>%
  group_by(system_label, year, month) %>%
  summarise(SUM = sum(miles_traveled), .groups = "keep") %>%
  ungroup() %>%
  group_by(system_label, year) %>%
  mutate(percent = SUM / sum(SUM)) %>%
  ungroup() %>%
  group_by(system_label, month) %>%
  summarise(
    MEAN = mean(percent),
    SE = sd(percent) / sqrt(n()), .groups = "keep"
  ) %>%
  mutate(label = if_else(month == "Dec", system_label, NA_character_)) %>%
  ggplot(aes(x = month, y = MEAN, group = system_label, color = system_label, pch = system_label)) +
  geom_segment(aes(x = 1, xend = 1, y = 0, yend = .04), lty = 3, color = "grey90") +
  geom_segment(aes(x = 4, xend = 4, y = 0, yend = .095), lty = 3, color = "grey90") +
  geom_segment(aes(x = 7, xend = 7, y = 0, yend = .17), lty = 3, color = "grey90") +
  geom_segment(aes(x = 10, xend = 10, y = 0, yend = .08), lty = 3, color = "grey90") +
  geom_point(size = 2) +
  geom_line() +
  scale_color_manual(values = system_label_cols[c("DNR State", "Greater MN Regional", "Metro Regional")]) +
  scale_y_continuous(labels = scales::label_percent(), expand = expansion(mult = 0, 0)) +
  theme_council_open(base_size = 10) +
  scale_x_discrete(
    expand = expansion(mult = c(0.01, 0.35))
  ) +
  coord_cartesian(clip = "off") +
  labs(
    x = "", caption = fig_caption,
    title = "Monthly trail use (% of annual trail miles traveled)"
  ) +
  geom_text(aes(x = 12.2, y = MEAN, label = label), hjust = 0, size = 3) +
  theme(
    legend.position = "none",
    axis.title.y = element_blank(),
    axis.title.x = element_blank(),
    plot.title = element_text(size = 10)
  )
# takeaway here: winter ns at GMN, double check if spring vs fall is different at metro
# car::Anova(lm(percent ~ system * year * as.character(month), data = filter(a, year != 2022)))

# ggsave(file.path(here::here(), "figures/storymap/seasonal-trail-use.png"))
```


### Hourly trail use varies by day and system

On weekends (Sa-Su), trails tend to have a late morning peak in visitation around 11 a.m. (Figure \@ref(fig:hourly-fig-trail)). On weekdays (M-F), trail visitation shows a bimodal use pattern. A peak in the late morning to noon time frame is still observed. A second peak in use occurs around 5 p.m. The first peak (morning) is larger for DNR state trails, while the second peak (evenings) is larger for regional trails. The strength of the second peak may reflect commuting patterns (i.e., trails being used as transportation to and from work) or the ease of accessing regional trails near home after work. 

```{r hourly-fig-trail}
#| fig.cap = "Daily trail use patterns during summer 2021 for each system. The total number of visits across all trails is summed for each system, from which the percent of total visits is calculated. Individual trail units may show trends that differ from the overall system trend.",

hourly_trails %>%
  mutate(day_type = str_remove_all(day_type, "1: Average |2: Average | Day")) %>%
  filter(!str_detect(day_type, "All")) %>%
  group_by(system_label, day_type, time) %>%
  summarise(total = sum(sumtotal), .groups = "keep") %>%
  ungroup() %>%
  group_by(system_label, day_type) %>%
  mutate(percent = total / sum(total)) %>%
  filter(!is.na(system_label)) %>%
  mutate(
    time2 = strftime(time, format = "%H:%M"),
    label = if_else(time2 == "23:00", system_label, NA_character_)
  ) %>%
  ggplot(aes(x = time, y = percent, color = system_label, group = system_label, pch = system_label)) +
  geom_point() +
  geom_line() +
  facet_wrap(~day_type, nrow = 1) +
  scale_color_manual(values = system_label_cols[c("DNR State", "Greater MN Regional", "Metro Regional")]) +
  theme_council_open(base_size = 10) +
  scale_y_continuous(labels = scales::label_percent()) +
  scale_x_datetime(date_breaks = "4 hours", date_labels = "%l%p", expand = expansion(mult = c(0.01, 0.01))) +
  labs(
    x = "", y = "System \ntotal", color = "", shape = "",
    title = "Hourly trail use (summer 2021)",
    caption = fig_caption
  ) +
  theme(
    legend.position = "bottom",
    panel.grid.major.x = element_line(linetype = 3),
    axis.title.y = element_blank(),
    axis.title.x = element_blank(),
    plot.title = element_text(size = 10)
  )

# ggsave(file.path(here::here(), "figures/storymap/hourly-trail-use.png"))
```


## Visitor modeshare varies by season 

During summer months, bicycle use generally makes up the majority of trail use (Figure \@ref(fig:trail-modeshare-fig)). However, this pattern switches in the winter when pedestrian use makes up the majority of use. Trails that are maintained for winter mountain or fat tire biking see a less pronounced decline in bike use during winter months. This is most common for DNR state and Greater Minnesota regional trails. 

We observe that this change is caused by a decrease in winter bicycle use, rather than by an increase in winter pedestrian use. It is likely that longer recreational bicycle trips decline in the winter, while shorter pedestrian trips in neighborhoods or other walkable areas remain more constant. 
 
```{r trail-modeshare-fig}
#| fig.cap = "Mode share of trips made to trails in summer (June - August) and winter seasons (December - February). Bars show the average annual mode share across the three complete years in this study (2019, 2020, 2021). Bar color indicates trail system. Individual trail units may show trends that differ from the overall system trend.",

monthly_trail_volume %>%
  mutate(season = if_else(month %in% c("Dec", "Jan", "Feb"), "Winter (December - February)", #
    if_else(month %in% c("Jun", "Jul", "Aug"), "Summer (June - August)", NA_character_)
  )) %>%
  filter(!is.na(season)) %>%
  group_by(system_label, season) %>%
  summarise(
    Bicycle = sum(bike_miles_traveled),
    Pedestrian = sum(ped_miles_traveled), .groups = "keep"
  ) %>%
  ungroup() %>%
  mutate(
    total = (Bicycle + Pedestrian),
    Bicycle = Bicycle / total,
    Pedestrian = Pedestrian / total
  ) %>%
  pivot_longer(
    cols = c(Bicycle, Pedestrian),
    names_to = "Mode",
    values_to = "Share"
  ) %>%
  ggplot(aes(x = Mode, y = Share, fill = system_label)) +
  geom_bar(stat = "identity", color = "black", position = position_dodge(width = .9)) +
  theme_council_open(base_size = 10) +
  scale_fill_manual(values = system_label_cols) +
  facet_wrap(~season) +
  scale_y_continuous(labels = scales::label_percent(), expand = expansion(mult = c(0.0, 0.05))) +
  labs(
    title = "Trail visitor travel mode, 2021",
    x = "", y = "", fill = "",
    caption = fig_caption
  ) +
  theme(
    legend.position = "bottom",
    axis.title.y = element_blank(),
    axis.title.x = element_blank(),
    plot.title = element_text(size = 10)
  ) +
  annotate("segment", x = -Inf, xend = -Inf, y = -Inf, yend = Inf, colour = "grey92", lwd = 1) +
  geom_label(aes(
    label = paste0(format(round(Share * 100, 0), nsmall = 0), "%"), group = system_label,
    y = rep(c(.1, .1, .1, .1, .1, .1), 2)
  ), fill = "white", size = 2.5, position = position_dodge(width = .9))


# ggsave(file.path(here::here(), "figures/storymap/trail-modeshare.png"))
```
