==> trails/06_clean_coverage_review_data.R <==
## This script cleans the data generated by 05_initiate_coverage_review. Raw LBS data is
## read from `data-intermediate/trails`. Home locations and demographics
## are processed for both summer 2021 and calendar year 2021. Processed data is saved
## to `data-intermediate/visitors`.

##### setup #####
load(file.path(here(), "data-intermediate/trails/coverage-review-data.rda"))
days_in_summer <- 91

##### process total activity #####
## total activity is needed to weight demographics & home locations by mode

#### summer 2021 ####
summer21_trails1 <- bind_rows(
  dnr_demos_homes$summer21_activity,
  gmn_demos_homes$summer21_activity,
  metro_demos_homes$summer21_activity
)

summer21_trails <- summer21_trails1 %>%
  mutate(start_date = "06/01/2021") %>%
  # join with cleaned trail segments so we can add 0 for segments dropped by StL
  full_join(
    st_drop_geometry(cleaned_trail_segments) %>%
      transmute(
        osm_id = osm_id,
        zone_name = zone_name,
        mode_of_travel = "Bicycle - StL Bicycle Volume",
        day_part = "0: All Day (12am-12am)",
        day_type = "0: All Days (M-Su)"
      ) %>%
      unique()
  ) %>%
  pivot_wider(names_from = mode_of_travel, values_from = average_daily_zone_traffic_st_l_volume) %>%
  rename(
    bike = `Bicycle - StL Bicycle Volume`,
    ped = `Pedestrian - StL Pedestrian Volume`
  ) %>%
  # convert to from average daily to total monthly volume
  mutate(
    bike = replace_na(bike, 0),
    ped = replace_na(ped, 0),
  ) %>%
  # get metadata with segment length
  left_join(trail_segment_metadata, by = "zone_name") %>%
  mutate(approx_length_miles = as.numeric(approx_length_miles)) %>%
  mutate(
    bike_use = bike * days_in_summer,
    ped_use = ped * days_in_summer,
    b_miles = bike * days_in_summer * approx_length_miles,
    p_miles = ped * days_in_summer * approx_length_miles
  ) %>%
  # summarise at segment level
  group_by(zone_name, day_type, day_part) %>%
  summarise(
    volume = sum(bike_use, ped_use),
    bike_use = sum(bike_use),
    ped_use = sum(ped_use),
    miles_used = sum(b_miles, p_miles),
    bike_miles = sum(b_miles),
    ped_miles = sum(p_miles),
    .groups = "keep"
  ) %>%
  # fill in missing values with 0
  ungroup() %>%
  # first, expand to identify missing values
  tidyr::complete(day_part, day_type, zone_name,
    fill = list(volume = 0, bike_use = 0, ped_use = 0, miles_used = 0, bike_miles = 0, ped_miles = 0)
  ) %>%
  mutate(
    modeshare_bike_use = bike_use / volume,
    modeshare_ped_use = ped_use / volume,
    modeshare_bike_miles = bike_miles / miles_used,
    modeshare_ped_miles = ped_miles / miles_used
  ) %>%
  # replace NaNs caused by 0/0 errors with 0 (keep NAs to indicate missing values)
  mutate(
    modeshare_bike_use = ifelse(volume == 0, NA, modeshare_bike_use),
    modeshare_ped_use = ifelse(volume == 0, NA, modeshare_ped_use),
    modeshare_bike_miles = ifelse(volume == 0, NA, modeshare_bike_miles),
    modeshare_ped_miles = ifelse(volume == 0, NA, modeshare_ped_miles)
  ) %>%
  # rejoin metadata
  left_join(trail_segment_metadata, by = "zone_name")

#### year 2021 ####
full21_trails1 <- bind_rows(
  dnr_demos_homes$full21_activity,
  gmn_demos_homes$full21_activity,
  metro_demos_homes$full21_activity
)

full21_trails <- full21_trails1 %>%
  mutate(start_date = "01/01/2021") %>%
  # join with cleaned trail segments so we can add 0 for segments dropped by StL
  full_join(
    st_drop_geometry(cleaned_trail_segments) %>%
      transmute(
        osm_id = osm_id,
        zone_name = zone_name,
        mode_of_travel = "Bicycle - StL Bicycle Volume",
        day_part = "0: All Day (12am-12am)",
        day_type = "0: All Days (M-Su)"
      ) %>%
      unique()
  ) %>%
  pivot_wider(names_from = mode_of_travel, values_from = average_daily_zone_traffic_st_l_volume) %>%
  rename(
    bike = `Bicycle - StL Bicycle Volume`,
    ped = `Pedestrian - StL Pedestrian Volume`
  ) %>%
  # convert to from average daily to total monthly volume
  mutate(
    bike = replace_na(bike, 0),
    ped = replace_na(ped, 0),
  ) %>%
  # get metadata with segment length
  left_join(trail_segment_metadata, by = "zone_name") %>%
  mutate(approx_length_miles = as.numeric(approx_length_miles)) %>%
  mutate(
    bike_use = bike * 365,
    ped_use = ped * 365,
    b_miles = bike * 365 * approx_length_miles,
    p_miles = ped * 365 * approx_length_miles
  ) %>%
  # summarise at segment level
  group_by(zone_name, day_type, day_part) %>%
  summarise(
    volume = sum(bike_use, ped_use),
    bike_use = sum(bike_use),
    ped_use = sum(ped_use),
    miles_used = sum(b_miles, p_miles),
    bike_miles = sum(b_miles),
    ped_miles = sum(p_miles),
    .groups = "keep"
  ) %>%
  # fill in missing values with 0
  ungroup() %>%
  # first, expand to identify missing values
  tidyr::complete(day_part, day_type, zone_name,
    fill = list(volume = 0, bike_use = 0, ped_use = 0, miles_used = 0, bike_miles = 0, ped_miles = 0)
  ) %>%
  mutate(
    modeshare_bike_use = bike_use / volume,
    modeshare_ped_use = ped_use / volume,
    modeshare_bike_miles = bike_miles / miles_used,
    modeshare_ped_miles = ped_miles / miles_used
  ) %>%
  # replace NaNs caused by 0/0 errors with 0 (keep NAs to indicate missing values)
  mutate(
    modeshare_bike_use = ifelse(volume == 0, NA, modeshare_bike_use),
    modeshare_ped_use = ifelse(volume == 0, NA, modeshare_ped_use),
    modeshare_bike_miles = ifelse(volume == 0, NA, modeshare_bike_miles),
    modeshare_ped_miles = ifelse(volume == 0, NA, modeshare_ped_miles)
  ) %>%
  # rejoin metadata
  left_join(trail_segment_metadata, by = "zone_name")

##### process demographics #####

#### summer 2021 ####
summer_trail_demographics1 <- bind_rows(
  dnr_demos_homes$summer21_demos,
  gmn_demos_homes$summer21_demos,
  metro_demos_homes$summer21_demos
)

summer_trail_demographics <- summer21_trails %>%
  select(-c(names(trail_segment_metadata %>% select(-zone_name, -unit_id)))) %>%
  # weight by bike/ped at segment level
  group_by(zone_name, day_type, day_part) %>%
  full_join(summer_trail_demographics1) %>%
  # use = bike miles or ped miles depending on mode, still at segment level
  mutate(use = ifelse(mode_of_travel == "Bicycle - StL Bicycle Volume", bike_miles, ped_miles)) %>%
  # transform percents to persons to weight bike & ped demographics
  mutate(across(white:graduate_or_professional_degree, ~ .x * use)) %>%
  select(-c(volume:modeshare_ped_miles, average_daily_zone_traffic_st_l_volume)) %>%
  # summarise at TRAIL level (previously we were missing a group by, I think)
  ungroup() %>%
  # get trail-level metadata
  left_join(trail_metadata, by = "unit_id") %>%
  group_by(unit_id, day_type, day_part) %>%
  summarise_if(is.numeric, sum, na.rm = TRUE) %>%
  ungroup() %>%
  mutate(across(white:graduate_or_professional_degree, ~ .x / use)) %>%
  # replace all NaNs with NAs
  mutate_all(~ ifelse(is.nan(.), NA, .))

#### full 2021 ####
full_trail_demographics1 <- bind_rows(
  dnr_demos_homes$full21_demos,
  gmn_demos_homes$full21_demos,
  metro_demos_homes$full21_demos
)

full_trail_demographics <- full21_trails %>%
  select(-c(names(trail_segment_metadata %>% select(-zone_name, -unit_id)))) %>%
  # weight by bike/ped at segment level
  group_by(zone_name, day_type, day_part) %>%
  full_join(full_trail_demographics1) %>%
  # use = bike miles or ped miles depending on mode, still at segment level
  mutate(use = ifelse(mode_of_travel == "Bicycle - StL Bicycle Volume", bike_miles, ped_miles)) %>%
  # transform percents to persons to weight bike & ped demographics
  mutate(across(white:graduate_or_professional_degree, ~ .x * use)) %>%
  select(-c(volume:modeshare_ped_miles, average_daily_zone_traffic_st_l_volume)) %>%
  # summarise at TRAIL level (previously we were missing a group by, I think)
  ungroup() %>%
  # get trail-level metadata
  left_join(trail_metadata, by = "unit_id") %>%
  group_by(unit_id, day_type, day_part) %>%
  summarise_if(is.numeric, sum, na.rm = TRUE) %>%
  ungroup() %>%
  mutate(across(white:graduate_or_professional_degree, ~ .x / use)) %>%
  # replace all NaNs with NAs
  mutate_all(~ ifelse(is.nan(.), NA, .))


##### save #####
save(summer21_trails, full21_trails,
  summer_trail_demographics, full_trail_demographics,
  file = file.path(here(), "data-intermediate/visitors/raw-trail-demographics.rda")
)
